backend/requirements.txt
fastapi
uvicorn
httpx
sentence-transformers
faiss-cpu
pydantic
python-dotenv
motor
python-multipart
